+++ src/lib/Thoughts.svelte
@@ -5,6 +5,15 @@
     let thoughts: IThought[] = $state([]);
     let text: string = $state("");
     let editingThoughtId: number | null = $state(null);
+    let searchedText: string = $state("");
+    let filteredThoughts = $derived(filterThoughts(searchedText));
+
+    function filterThoughts(query: string) {
+        if (!query || !query.trim()) return thoughts;
+
+        const lower = query.toLowerCase();
+        return thoughts.filter((a) => a.text.toLowerCase().includes(lower));
+    }
 
     onMount(() => {
         // Load thoughts from local storage or an API
@@ -27,8 +36,10 @@
     }
 
     function handleDelete(id: number) {
-        thoughts = thoughts.filter((thought) => thought.id !== id);
-        localStorage.setItem("thoughts", JSON.stringify(thoughts));
+        if (confirm("Are you sure you want to delete this thought?")) {
+            thoughts = thoughts.filter((thought) => thought.id !== id);
+            localStorage.setItem("thoughts", JSON.stringify(thoughts));
+        }
     }
 
     function handleEdit(id: number) {
@@ -40,64 +51,77 @@
     }
 </script>
 
-<div class="space-y-2">
-    <textarea
-        bind:value={text}
-        class="textarea h-24 w-full"
-        placeholder="Thought"
-        onkeydown={(e) => {
-            if (e.key == "Enter") {
-                handleSave();
-            }
-        }}
-    ></textarea>
-    <button
-        class="btn btn-primary"
-        onclick={() => {
-            if (editingThoughtId) {
-                // Update existing thought
-                thoughts = thoughts.map((thought) =>
-                    thought.id === editingThoughtId
-                        ? { ...thought, text }
-                        : thought,
-                );
-                editingThoughtId = null;
-                localStorage.setItem("thoughts", JSON.stringify(thoughts));
-                text = "";
-            } else {
-                // Save new thought
-                handleSave();
-            }
-        }}
-        disabled={!text}
-    >
-        Save
-    </button>
-</div>
+<div class="space-y-5">
+    <div class="space-y-2">
+        <textarea
+            bind:value={text}
+            class="textarea h-24 w-full"
+            placeholder="Thought"
+            onkeydown={(e) => {
+                if (e.key == "Enter") {
+                    handleSave();
+                }
+            }}
+        ></textarea>
+        <button
+            class="btn btn-primary"
+            onclick={() => {
+                if (editingThoughtId) {
+                    // Update existing thought
+                    thoughts = thoughts.map((thought) =>
+                        thought.id === editingThoughtId
+                            ? { ...thought, text }
+                            : thought,
+                    );
+                    editingThoughtId = null;
+                    localStorage.setItem("thoughts", JSON.stringify(thoughts));
+                    text = "";
+                } else {
+                    // Save new thought
+                    handleSave();
+                }
+            }}
+            disabled={!text}
+        >
+            Save
+        </button>
+    </div>
 
-<div class="space-y-2 mt-10">
-    {#each thoughts as thought (thought.id)}
-        <div class="card bg-base-300 shadow-md">
-            <div class="card-body">
-                <p>{thought.text}</p>
-                <!-- date, edit, delete -->
-                <div class="flex justify-between items-center">
-                    <p class="text-xs text-gray-500">
-                        {new Date(thought.createdAt).toLocaleString()}
-                    </p>
+    <input
+        type="search"
+        placeholder="Search thoughts..."
+        class="input w-full"
+        bind:value={searchedText}
+    />
 
-                    <div class="flex gap-2">
-                        <button
-                            class="btn btn-sm"
-                            onclick={() => handleEdit(thought.id)}>Edit</button
-                        >
-                        <button
-                            class="btn btn-sm btn-error"
-                            onclick={() => handleDelete(thought.id)}>X</button
-                        >
+    <div class="space-y-2">
+        {#each filteredThoughts as thought (thought.id)}
+            <div class="card bg-base-300 shadow-md">
+                <div class="card-body">
+                    <p>{thought.text}</p>
+                    <!-- date, edit, delete -->
+                    <div class="flex justify-between items-center">
+                        <p class="text-xs text-gray-500">
+                            {new Date(thought.createdAt).toDateString()}
+                            -
+                            {new Date(thought.createdAt).toLocaleTimeString()}
+                        </p>
+
+                        <div class="flex gap-2">
+                            <button
+                                class="btn btn-sm"
+                                onclick={() => handleEdit(thought.id)}
+                                >Edit</button
+                            >
+                            <button
+                                class="btn btn-sm btn-error"
+                                onclick={() => handleDelete(thought.id)}
+                                >X</button
+                            >
+                        </div>
                     </div>
                 </div>
             </div>
-        </div>
-    {/each}
+        {/each}
+    </div>
 </div>
